KNOX-1997: Refactoring test cases
