--- gateway-server/src/test/java/org/apache/knox/gateway/websockets/WebsocketServerInitiatedEchoHandler.java
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.knox.gateway.websockets;
-
-import org.eclipse.jetty.websocket.server.WebSocketHandler;
-import org.eclipse.jetty.websocket.servlet.ServletUpgradeRequest;
-import org.eclipse.jetty.websocket.servlet.ServletUpgradeResponse;
-import org.eclipse.jetty.websocket.servlet.WebSocketCreator;
-import org.eclipse.jetty.websocket.servlet.WebSocketServletFactory;
-
-/**
- * A Mock websocket handler that just Echos messages
- *
- */
-public class WebsocketServerInitiatedEchoHandler extends WebSocketHandler implements WebSocketCreator {
-  private final ServerInitiatingMessageSocket socket = new ServerInitiatingMessageSocket();
-
-  @Override
-  public void configure(WebSocketServletFactory factory) {
-      factory.getPolicy().setMaxTextMessageSize(2 * 1024 * 1024);
-      factory.setCreator(this);
-  }
-
-  @Override
-  public Object createWebSocket(ServletUpgradeRequest req, ServletUpgradeResponse resp) {
-      return socket;
-  }
-}
